backend/requirements.txt
# FastAPI and ASGI Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic and Settings
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# SQLAlchemy and Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Security and Encryption
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0
bcrypt==4.1.2

# Rate Limiting
slowapi>=0.1.9

# Supabase Client
supabase==2.3.0

# HTTP Client (version compatible with supabase)
httpx>=0.24.0,<0.25.0

# Caching
cachetools==5.3.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Development Dependencies (optional, comment out for production)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==23.12.1
ruff==0.1.11
mypy==1.8.0

# Optional: For better async performance
# uvloop==0.19.0

# Document Parsing (LlamaParse for PDF bank statements)
llama-cloud-services>=0.6.0

# Fuzzy Matching (for bank reconciliation)
python-Levenshtein==0.25.1
thefuzz==0.22.1

# PDF Generation
weasyprint>=60.0
Jinja2==3.1.3

# Email Service
aiosmtplib>=3.0.0
email-validator>=2.1.0
